fix :: 테이블의 데이터타입(BIGINT) 와 엔티티의 타입(Integer) 가 서로 맞지 않음 -> Long 으로 수정

+++ module-domain/src/main/java/com/kernel360/commoncode/entity/CommonCode.java
@@ -23,7 +23,7 @@ public class CommonCode extends BaseEntity {
     private String codeName;
 
     @Column(name = "upper_no")
-    private Integer upperNo;
+    private Long upperNo;
 
     @Column(name = "upper_name", length = Integer.MAX_VALUE)
     private String upperName;
